requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-import-export==4.1.1
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
